OML Add osmocom-specific way to deactivate radio link timeout
In some situations (e.g. when trying to do measurements/testing on the
BTS receiver / uplink) it is useful to have a way to disable the radio
link timeout and keep any channel open until deactivated, irrespective
of whether (valid) data is received or not.

This adds a related feature that can be activated by using an
osmocom-specific value of 0xff for the TS 12.21 Connection Failure
Criterion (9.4.14).

+++ include/osmo-bts/gsm_data.h
@@ -94,7 +94,9 @@ struct gsm_bts_role_bts {
 		uint8_t tc4_ctr;
 	} si;
 	struct gsm_time gsm_time;
-	uint8_t radio_link_timeout;
+	/* Radio Link Timeout counter. -1 disables timeout for
+	 * lab/measurement purpose */
+	int radio_link_timeout;
 
 	int ul_power_target;		/* Uplink Rx power target */
